wip(agents): start installer/auth/readiness file regroup
Scaffolding moves for the agents domain regroup (per
guides/mental-model.md): install_lock -> installer/lock, managed_npm ->
installer/managed_npm, installer.rs split into installer/{mod,service,
tests}, login_terminal -> auth_config/, auth_config test files
un-#[path]-hacked, readiness/resolver -> readiness/service. Does not
compile yet — module rewiring lands in the next commits.
