+++ anyharness/crates/anyharness-lib/src/domains/agents/auth_config/claude_tests.rs
