+++ anyharness/crates/anyharness-lib/src/domains/agents/auth_config/scope_tests.rs
