+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/lock.rs
