+++ anyharness/crates/anyharness-lib/src/domains/agents/installer/managed_npm.rs
